Make sure value implements IConvertible before attempting to use Convert.ChangeType() (#35)
When returning values, conversion was attempted using the IConvertible interface, but without first checking if the value implemented IConvertible - leading to an exception in the case(s) where IConvertible was not implemented.

This change adds a test around the conversion attempt, to ensure conversion by IConvertible is only attempted on values that implement IConvertible.